Connect dc filters to table
